New ReadME.md file with detailed information about the project
